integrated all the tree functions into Simulate
create a maximum_plots, random_plots, dijikstra_plots folder for seperate graphs.
run 1 function at a time in simulate for each graph and make sure to change last parameter of visualize_Tree.
also made it so that whenever program is run for some algo it deletes previous graphs in the folder for that algo